Fix Ctrl+W on Dvorak-QwertyCmd layouts.
It seems that this layout will cause the input method methods to be
called, whereas "normal" layouts don't go this route. The problem is
that this sends the wrong text (namely a ',' in the case when the user
pressed 'w'). However, we also need to set the
"interpretKeyEventsSwallowedKey flag since in this particular case we
really mean that a selector didn't swallow this key.

Yes, it's complicated and this layout is used by a very small minority.
But this layout is handy enough to support, kind of like the Mac users
of old. :-)