Suppress warnings when compiling makeicns utility
A few NSFileManager and CG methods were deprecated in OS X v10.5 and are
now causing warnings during compilation. Because of this the custom
icons have been made optional and the default build does not provide
them anymore.

This is a workaround fix to suppress the warning messages until the
source code is updated properly.