Fix getfontname().
getfontname() would incorrectly return strings of the form
"FONTNAME:SIZE", when it should be returning strings of the form
"FONTNAME:hSIZE"--notice the "h" in the last form.

To fix this, we change the internal representation of the font name and
size to include the "h", and peel it off when actually setting the font.